wysiwyl/requirements.txt
anyascii==0.1.7
asgiref==3.3.1
beautifulsoup4==4.8.2
certifi==2020.12.5
chardet==4.0.0
deap==1.3.1
Django==3.1.7
django-debug-toolbar==3.2
django-extensions==3.1.1
django-filter==2.4.0
django-modelcluster==5.1
django-taggit==1.3.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
html5lib==1.1
idna==2.10
joblib==1.0.1
l18n==2020.6.1
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.3
Pillow==8.1.2
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
stopit==1.1.2
tablib==3.0.0
threadpoolctl==2.1.0
TPOT==0.11.7
tqdm==4.59.0
update-checker==0.18.0
urllib3==1.26.4
wagtail==2.12.3
wagtail-modelchooser==2.3.3
webencodings==0.5.1
Willow==1.4
xgboost==1.3.3
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
